feat: adde deltaX fo fuselage calculatio
refactor: many things
