+++ frontend/src/pages/stabillity/longitudonalMoment/StabillityLongitudalMomentChart.tsx
@@ -10,8 +10,7 @@ const StabillityLongitudalMomentChart = () => {
   const alfa = useLongitudalMomentStore((state) => state.alfa);
   const cmbu = useLongitudalMomentOutput((state) => state.cmbu);
 
-  const traceAlfa = { x: alfa, y: cmbu };
-  const traceCz = { x: cz, y: cmbu };
+
   const layouts = (type: string) => {
     let layout = {
       tittle:{text:"Longitudonal Moment"}, 
@@ -30,10 +29,11 @@ const StabillityLongitudalMomentChart = () => {
   };
 
   useEffect(()=>{
-    console.log(traceAlfa)
+    const traceAlfa = { x: alfa, y: cmbu };
+    const traceCz = { x: cz, y: cmbu };
     Plotly.newPlot("Cmbu_alfa",[traceAlfa],layouts("alfa") )
     Plotly.newPlot("Cmbu_cz",[traceCz],layouts("cz") )
-  },[traceAlfa,traceCz ])
+  },[alfa,cz,cmbu ])
   return <div>
     <div id="Cmbu_alfa"/>
     <div id ="Cmbu_cz"/>
